python/requirements.txt
absl-py==0.7.0
aiohttp==3.5.4
astor==0.7.1
atomicwrites==1.3.0
attrs==18.2.0
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
git+https://github.com/thoughtworksarts/EmoPy.git
Flask==1.0.2
gast==0.2.2
graphviz==0.10.1
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Lasagne==0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
more-itertools==6.0.0
networkx==2.2
numpy==1.17.4
opencv-python==4.0.0.21
Pillow==6.2.2
pluggy==0.8.1
protobuf==3.6.1
py==1.8.0
pybase64==0.5.0
pydot==1.4.1
pyparsing==2.3.1
pytest==4.3.0
python-dateutil==2.8.0
python-twitter==3.5
PyWavelets==1.0.1
PyYAML==4.2b1
scikit-image==0.14.2
scikit-learn==0.20.2
scikit-neuralnetwork==0.7
scipy==1.0.0
six==1.12.0
tensorboard==1.13.0
tensorflow==1.15.2
termcolor==1.1.0
Theano==1.0.4
toolz==0.9.0
Werkzeug==0.15.3
